+++ CHANGELOG.rst
@@ -7,6 +7,32 @@ Change Log
 ----------
 
 
+3.10.0
+======
+
+* Adds support for ``creds_utils``.
+
+  * Class ``KeyManager``, with methods:
+
+    * ``KeyManager.get_keydict_for_env(self, env)``
+
+    * ``KeyManager.get_keydict_for_server(self, server)``
+
+    * ``KeyManager.get_keydicts(self)``
+
+    * ``KeyManager.get_keypair_for_env(self, env)``
+
+    * ``KeyManager.get_keypair_for_server(self, server)``
+
+    * ``KeyManager.keydict_to_keypair(auth_dict)``
+
+    * ``KeyManager.keypair_to_keydict(auth_tuple, *, server)``
+
+  * Class ``FourfrontKeyManager``
+
+  * Class ``CGAPKeyManager``
+
+
 3.9.0
 =====